+++ weblate/checks/chars.py
@@ -17,24 +17,14 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
+import re
+
 from django.utils.translation import gettext_lazy as _
 
 from weblate.checks.base import CountingCheck, TargetCheck, TargetCheckParametrized
 from weblate.checks.markup import strip_entities
 from weblate.checks.parser import single_value_flag
 
-KASHIDA_CHARS = (
-    "\u0640",
-    "\uFCF2",
-    "\uFCF3",
-    "\uFCF4",
-    "\uFE71",
-    "\uFE77",
-    "\uFE79",
-    "\uFE7B",
-    "\uFE7D",
-    "\uFE7F",
-)
 FRENCH_PUNCTUATION = {";", ":", "?", "!"}
 
 
@@ -383,11 +373,19 @@ class KashidaCheck(TargetCheck):
     name = _("Kashida letter used")
     description = _("The decorative kashida letters should not be used")
 
+    kashida_regex = (
+        # Allow kashida after certain letters
+        "(?<![\u0628\u0643\u0644])"
+        # List of kashida letters to check
+        "[\u0640\uFCF2\uFCF3\uFCF4\uFE71\uFE77\uFE79\uFE7B\uFE7D\uFE7F]"
+    )
+    kashida_re = re.compile(kashida_regex)
+
     def check_single(self, source, target, unit):
-        return any(x in target for x in KASHIDA_CHARS)
+        return self.kashida_re.search(target)
 
     def get_fixup(self, unit):
-        return [("[{}]".format("".join(KASHIDA_CHARS)), "", "gu")]
+        return [(self.kashida_regex, "", "gu")]
 
 
 class PunctuationSpacingCheck(TargetCheck):