+++ weblate/checks/tests/test_chars_checks.py
@@ -340,6 +340,7 @@ class KashidaCheckTest(CheckTestCase):
     def setUp(self):
         super().setUp()
         self.test_good_matching = ("string", "string", "")
+        self.test_good_ignore = ("string", "بـ:", "")
         self.test_failure_1 = ("string", "string\u0640", "")
         self.test_failure_2 = ("string", "string\uFE79", "")
         self.test_failure_3 = ("string", "string\uFE7F", "")