+++ docs/changes.rst
@@ -8,6 +8,7 @@ Not yet released.
 * Fixed API display of branch in components with a shared repository.
 * Improved analysis on the failed push alert.
 * Fixed manually editing page when browsing changes.
+* Improved accuracy of Kashida check.
 
 Weblate 4.10
 ------------